requirements.txt
pywinrm==0.4.2
